feat: extract weather and BBS logic into separate modules
